mobileotp/requirements.txt
Django==3.2
Pillow==5.1.0
gunicorn==20.1.0
psycopg2-binary==2.8.6
africastalking==1.2.3
django-cors-headers==3.4.0
django-filter==2.3.0
django-js-asset==1.2.2
django-subadmin==2.0.1
djangorestframework==3.12.2
drf-extra-fields==3.0.3
django-leaflet==0.27.1
# Devices requesting services
pyyaml
ua-parser
user-agents
django-user-agents